requirements.txt
fastapi
uvicorn[standard]
pydantic
google-cloud-bigquery
psycopg2-binary
SQLAlchemy
python-dotenv
